refactor: 내부에서 인스턴스 생성 방식으로 변경
